requirements.txt
github.py == 0.5.0